+++ lib/mayaUsd/ufe/UsdPathMappingHandler.cpp
@@ -138,7 +138,7 @@ Ufe::Path UsdPathMappingHandler::fromHost(const Ufe::Path& hostPath) const
         mayaComps.emplace_back(mayaHostPath.back());
         mayaHostPath = mayaHostPath.pop();
         Ufe::Path ufePath;
-        if (PXR_NS::PrimUpdaterManager::readPullInformation(dagPath, ufePath)) {
+        if (PXR_NS::readPullInformation(dagPath, ufePath)) {
             // From the pulled info path, we pop only the last component and
             // append the Maya component array.
             std::reverse(mayaComps.begin(), mayaComps.end());
